Refactored fund-release API
